backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0.post1
edge-tts==7.2.3
pydub==0.25.1
python-multipart==0.0.6
httpx==0.25.2
python-dotenv==1.0.0
pyttsx3==2.90
ffmpeg-python==0.2.0
sqlalchemy==2.0.23
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
psycopg2-binary==2.9.9
asyncpg==0.29.0